Исправил проблемы отображения, добавил функцию раскрытия точки маршрута, домашнее задание 3.1
