+++ src/main.js
@@ -1,5 +1,4 @@
 import {generateEvents} from "./mock/generate-trip-events";
-import TripDaysContainerView from "./view/trip-days-container";
 import HeaderTripCostView from "./view/header-trip-cost";
 import HeaderNavMenuView from "./view/header-nav-menu";
 import HeaderFiltersView from "./view/header-filters";
@@ -9,10 +8,33 @@ import TripInfoContainerView from "./view/trip-info-container";
 import TripInfoRouteView from "./view/trip-info-route";
 import RenderEventFormView from "./view/render-event-form";
 import TripEventView from "./view/render-trip-event";
-import {getRandomInteger, renderElement} from "./utils";
+import {renderElement} from "./utils";
 
 const RENDER_EVENTS_COUNT = 4;
 const events = generateEvents(RENDER_EVENTS_COUNT);
+
+// Обработчики события смены формы на список и наоборот
+
+export const addEventToList = (eventListElement, event) => {
+  const tripEvent = new TripEventView(event);
+  const eventSwitchButton = tripEvent.getElement().querySelector(`.event__rollup-btn`);
+  const tripForm = new RenderEventFormView(event, event.counter);
+  const eventEditForm = tripForm.getElement();
+
+  const eventSwitchButtonHandler = () => {
+    eventListElement.replaceChild(tripForm.getElement(), tripEvent.getElement());
+  };
+
+  const formSwitchSubmitHandler = (evt) => {
+    evt.preventDefault();
+    eventListElement.replaceChild(tripEvent.getElement(), tripForm.getElement());
+  };
+
+  eventSwitchButton.addEventListener(`click`, eventSwitchButtonHandler);
+  eventEditForm.addEventListener(`submit`, formSwitchSubmitHandler);
+
+  renderElement(eventListElement, tripEvent.getElement());
+};
 
 // шапка
 
@@ -36,7 +58,6 @@ renderElement(tripControlsSecondHeaderElement, new HeaderFiltersView().getElemen
 // сортировка
 
 renderElement(tripEventsElement, new TripSortingView().getElement());
-renderElement(tripEventsElement, new TripDaysContainerView().getElement());
 
 const tripSortElement = tripEventsElement.querySelector(`.trip-sort`);
 
@@ -45,10 +66,11 @@ renderElement(tripSortElement, new TripEventsContainerView().getElement(), `inse
 const eventsListElement = tripEventsElement.querySelector(`.trip-events__list`);
 
 // форма
-renderElement(eventsListElement, new RenderEventFormView(events[getRandomInteger(0, events.length)], events.counter).getElement(),`insertBefore`);
+// renderElement(eventsListElement, new RenderEventFormView(events[getRandomInteger(0, events.length)], events.counter).getElement(),`insertBefore`);
 
 // Создание моков
 
 events.forEach((event) => {
-  renderElement(eventsListElement, new TripEventView(event).getElement(),);
+  addEventToList(eventsListElement, event);
+// renderElement(eventsListElement, new TripEventView(event).getElement());
 });
